GenAI/Frameworks/LlamaIndex/Information-Retriever-Using-llamaIndex/requirements.txt
llama-index
google-generativeai
llama-index-llms-gemini
pypdf
python-dotenv
IPython
streamlit
llama-index-llms-google-genai
llama-index-vector-stores-postgres
pinecone
llama_index.vector_stores.pinecone
llama-index-llms-google-genai
llama-index-embeddings-google-genai

#-e .